refactor: move version utils to util package
