+++ cmd/tidb-dashboard/main.go
@@ -47,8 +47,8 @@ import (
 	keyvisualregion "github.com/pingcap/tidb-dashboard/pkg/keyvisual/region"
 	"github.com/pingcap/tidb-dashboard/pkg/swaggerserver"
 	"github.com/pingcap/tidb-dashboard/pkg/uiserver"
-	"github.com/pingcap/tidb-dashboard/pkg/utils/version"
 	_ "github.com/pingcap/tidb-dashboard/populate/distro"
+	"github.com/pingcap/tidb-dashboard/util/versionutil"
 )
 
 type DashboardCLIConfig struct {
@@ -95,7 +95,7 @@ func NewCLIConfig() *DashboardCLIConfig {
 
 	flag.Parse()
 	if *showVersion {
-		version.PrintStandaloneModeInfo()
+		versionutil.PrintStandaloneModeInfo()
 		_ = log.Sync()
 		os.Exit(0)
 	}
